requirements.txt
Flask
Flask-Dance
Flask-SQLAlchemy

requests
python-dotenv
python-slugify
